+++ src/main/java/com/ururulab/ururu/order/dto/request/CartItemAddRequest.java
@@ -1,7 +1,7 @@
-package com.ururulab.ururu.order.controller.dto.request;
+package com.ururulab.ururu.order.dto.request;
 
-import com.ururulab.ururu.order.controller.dto.validation.CartValidationConstants;
-import com.ururulab.ururu.order.controller.dto.validation.CartValidationMessages;
+import com.ururulab.ururu.order.dto.validation.CartValidationConstants;
+import com.ururulab.ururu.order.dto.validation.CartValidationMessages;
 import jakarta.validation.constraints.Max;
 import jakarta.validation.constraints.Min;
 import jakarta.validation.constraints.NotNull;
