+++ src/main/java/com/ururulab/ururu/order/dto/validation/OrderValidationMessages.java
@@ -1,4 +1,4 @@
-package com.ururulab.ururu.order.controller.dto.validation;
+package com.ururulab.ururu.order.dto.validation;
 
 import lombok.experimental.UtilityClass;
